Split CLI check and emit commands
