+++ src/cli.rs
@@ -3,6 +3,7 @@ use std::process;
 
 mod build_graph_execution;
 mod build_graph_loading;
+mod check_emit_commands;
 mod compile;
 mod diagnostics;
 mod execution_commands;
@@ -15,6 +16,7 @@ use build_graph_execution::{
     test_build_targets, validate_build_graph_sources,
 };
 use build_graph_loading::load_build_graph;
+use check_emit_commands::{cmd_check, cmd_emit};
 use compile::{compile_file_to_binary, compile_file_to_c_source, typed_program_to_c_source};
 use diagnostics::{print_diagnostic, print_errors};
 use execution_commands::{cmd_build, cmd_build_graph, cmd_run_file, cmd_test};
@@ -96,36 +98,6 @@ pub fn main() {
             process::exit(1);
         }
     }
-}
-
-fn cmd_check(path_str: &str) {
-    if is_build_zen_path(path_str) {
-        let graph = load_build_graph(path_str);
-        let build_path = Path::new(path_str);
-        let base_dir = build_path.parent().unwrap_or_else(|| Path::new("."));
-        validate_build_graph_sources(base_dir, &graph);
-        check_build_graph_sources(base_dir, &graph);
-        println!("  {} build targets — ok", graph.targets().len());
-        return;
-    }
-
-    let typed = graph_frontend(path_str);
-    println!(
-        "  {} functions, {} types — ok",
-        typed.functions.len(),
-        typed.types.len()
-    );
-}
-
-fn cmd_emit(path_str: &str) {
-    if is_build_zen_path(path_str) {
-        let target = single_executable_build_target(path_str);
-        print!("{}", compile_file_to_c_source(&target.root_path));
-        return;
-    }
-
-    let typed = graph_frontend(path_str);
-    print!("{}", typed_program_to_c_source(&typed));
 }
 
 fn is_build_zen_path(path_str: &str) -> bool {
